requirements.txt
PyQt6>=6.4.0
cryptography>=41.0.0
keyring>=24.0.0
pywin32>=306; sys_platform == 'win32'
python-xlib>=0.33; sys_platform == 'linux'
